faat(expense): adding the buisness logic on the expense updateById

+++ src/services/expense-service.js
@@ -1,10 +1,19 @@
 const SequelizeErrorWrapper = require("../helpers/sequelize-error-wrapper");
 const Expense = require("../models").Expense;
-const { expenseSchema } = require("../validators/expense-validator");
-const { validateUserId } = require("../validators/user-validator");
+const {
+  handleAccountBalanceError,
+  handleIsPaidChange,
+} = require("../helpers/expense-helpers");
+const {
+  expenseSchema,
+  validatePresentExpenseProps,
+} = require("../validators/expense-validator");
 const { ApiValidationError } = require("../errors/validation-errors");
 const errorConstants = require("../constants/error-constants");
 const TransactionService = require("../services/transaction-service");
+const TransactionTypesEnum = require("../enums/transaction-types-enum");
+const { ApiInvalidArgumentError } = require("../errors/argument-errors");
+const ErrorMessageFormatter = require("../utils/error-message-formatter");
 
 /**
  * Class responsible for providing expense-related services.
@@ -17,7 +26,10 @@ class ExpenseService {
    * @returns {Promise<Array>} A Promise that resolves to an array with all the user's expenses.
    */
   static async getAll(userId) {
-    await validateUserId(userId);
+    if (!userId)
+      throw new ApiInvalidArgumentError(
+        ErrorMessageFormatter.missingArgument("userId"),
+      );
 
     return await Expense.findAll({
       where: {
@@ -34,7 +46,10 @@ class ExpenseService {
    * @returns {Promise<Object>} A Promise that resolves to the found expense object.
    */
   static async getById(id, userId) {
-    await validateUserId(userId);
+    if (!userId)
+      throw new ApiInvalidArgumentError(
+        ErrorMessageFormatter.missingArgument("userId"),
+      );
 
     return await Expense.findOne({
       where: {
@@ -52,14 +67,20 @@ class ExpenseService {
    * @returns {Promise<Object>} A Promise that resolves to the created expense object.
    */
   static async create(expense, userId) {
+    if (!userId)
+      throw new ApiInvalidArgumentError(
+        ErrorMessageFormatter.missingArgument("userId"),
+      );
+
     expense = { ...expense, userId: userId };
 
     await expenseSchema
+      .strict()
       .validate(expense, { abortEarly: false })
       .catch((err) => {
         throw new ApiValidationError(
           errorConstants.MSG_VALIDATION_ERROR,
-          err.message,
+          err.errors,
         );
       });
 
@@ -68,27 +89,14 @@ class ExpenseService {
     );
 
     if (createdExpense.isPaid) {
-      return await TransactionService.createFromExpense(createdExpense)
-        .then((createdTransaction) => {
-          return {
-            createdExpense: createdExpense,
-            accountTransaction: createdTransaction,
-          };
-        })
-        .catch(async (err) => {
-          /**
-           * If the amount overpass the account overdraftLimit this exception will occours
-           * in this way the expense will be set as unpaid but still created.
-           */
-          if (err.name === "AccountBalanceError") {
-            createdExpense.isPaid = false;
-            await createdExpense.save();
-            return {
-              createdExpense: createdExpense,
-              info: err.message,
-            };
-          } else throw err;
-        });
+      await TransactionService.createFromExpense(
+        createdExpense,
+        TransactionTypesEnum.WITHDRAWL,
+      ).catch(async (err) => {
+        if (err.name === "AccountBalanceError") {
+          return await handleAccountBalanceError(err, createdExpense);
+        } else throw err;
+      });
     }
 
     return createdExpense;
@@ -103,29 +111,41 @@ class ExpenseService {
    * @returns {Promise<Object>} A Promise that resolves to the updated expense object.
    */
   static async updateById(expense, id, userId) {
-    await expenseSchema
-      .validate(expense, { abortEarly: false, stripUnknown: true })
-      .catch((err) => {
-        throw new ApiValidationError(
-          errorConstants.MSG_VALIDATION_ERROR,
-          err.message,
-        );
-      });
+    if (!userId)
+      throw new ApiInvalidArgumentError(
+        ErrorMessageFormatter.missingArgument("userId"),
+      );
 
-    const dbTransaction = sequelize.transaction();
+    expense = { ...expense, userId: userId };
 
-    const updatedExpense = await Expense.update(expense, {
+    const errors = await validatePresentExpenseProps(expense);
+    if (!errors.isEmpty())
+      throw new ApiValidationError(
+        errorConstants.MSG_VALIDATION_ERROR,
+        errors.getErrors(),
+      );
+
+    const expenseBeforeUpdt = await this.getById(id, userId);
+
+    await Expense.update(expense, {
       where: {
         id: Number(id),
         userId: Number(userId),
       },
-    })
-      .then(async () => {
-        return await this.getById(id, userId);
-      })
-      .catch((err) => {
-        SequelizeErrorWrapper.wrapError(err);
-      });
+    }).catch((err) => {
+      SequelizeErrorWrapper.wrapError(err);
+    });
+
+    const updatedExpense = await this.getById(id, userId);
+
+    const hasChangedIsPaidField =
+      expenseBeforeUpdt.isPaid !== updatedExpense.isPaid;
+
+    if (hasChangedIsPaidField) {
+      await handleIsPaidChange(updatedExpense);
+    }
+
+    return updatedExpense;
   }
 
   /**
@@ -136,7 +156,10 @@ class ExpenseService {
    * @returns {Promise<number>} A Promise that resolves with the number of deleted rows (0 or 1).
    */
   static async deleteById(id, userId) {
-    await validateUserId(userId);
+    if (!userId)
+      throw new ApiInvalidArgumentError(
+        ErrorMessageFormatter.missingArgument("userId"),
+      );
 
     return await Expense.destroy({
       where: {
